Add manifest validation to common workflow repository
We run a manifest validation script as part of the
ocp-on-nerc/nerc-ocp-config repository. We would like to both enhance the
validations performed by that script and expose those features to other
repositories.

This commit adds the script to our common workflow repository and adds
support for kubeconform [1] for manifest validation.

[1]: https://github.com/yannh/kubeconform