Use new schema repository layout
In order to resolve validation failures on ocp-on-nerc/nerc-ocp-config#609,
we need to update the schema repository. This required some changes to the
repository layout because the tool on which we had been relying was no
longer maintained.

This commit updates our validation workflow to use the new schema
repository structure.